pyproject.toml
[tool.poetry]
name = "test-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Maksim <tolstikov_m@mail.ru>"]

[tool.poetry.dependencies]
python = "3.9.5"
bottle = "^0.12.19"
requests = "^2.26.0"
python-dotenv = "^0.18.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
